Prepare to remove barevisor

+++ src/win_hv/src/services/process_services.rs
@@ -99,7 +99,7 @@ pub(crate) fn get_process_field_sync(request: GetProcessFieldRequest) -> Hypervi
         ProcessField::NtPath(_) => {
             let field = process.get_nt_path();
             let raw_string = field.get_raw_bytes();
-            let offset = state.write_result(raw_string.as_ptr(), raw_string.len());
+            let offset = state.write_result(raw_string.as_ptr(), raw_string.len() as u32);
             ProcessField::NtPath(offset)
         }
         ProcessField::Protection(_) => ProcessField::Protection(process.get_protection()),
@@ -110,7 +110,7 @@ pub(crate) fn get_process_field_sync(request: GetProcessFieldRequest) -> Hypervi
         ProcessField::Token(_) => ProcessField::Token(process.get_token() as _),
         ProcessField::Threads(_) => {
             let thread_numbers = process.get_threads();
-            let offset = state.write_result(thread_numbers.as_ptr() as _, thread_numbers.len());
+            let offset = state.write_result(thread_numbers.as_ptr() as _, thread_numbers.len() as u32);
             ProcessField::Threads(offset)
         }
         ProcessField::DirectoryTableBase(_) => {
